+++ spec/internal/app/interactors/kafka_producer/outbox_transport_factory.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module KafkaProducer
+  class OutboxTransportFactory
+    def self.build(topic:, kafka: {})
+      ::Producer.new("outbox_item_topic", kafka)
+    end
+  end
+end