+++ spec/internal/app/models/inbox_item.rb
@@ -2,10 +2,4 @@
 
 class InboxItem < Sbmt::Outbox::InboxItem
   validates :event_name, :proto_payload, presence: true
-
-  def transports
-    [
-      ImportOrder
-    ]
-  end
 end