+++ CHANGELOG.md
@@ -11,9 +11,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
-- Drop supporting of Rails 5.0
-
 ### Fixed
+
+## [4.0.1] - 2023-04-07
+
+### Changed
+
+- Drop supporting of Rails 5.0
+- Add `transports` config option
+
 
 ## [4.0.1] - 2023-04-07